chore: Disable OpenSSH deamon
Previously used to tunnel traffic through for Firefox. Used this
approach to sever the connection from Firefox to the web, without
disconnecting the machine from the internet. Haven't thought about
another way of doing this, but since I haven't used SSH tunnels for this
use case in a while, it makes no sense keeping this port open :wink:.