chore: Remove emacs-overlay from system config
User should manage emacs-overlay from their home-configuration since
system-wide overlays are not exposed to nix-env or other nix tooling.
This complicates reasoning through a system configuration and the impact
of other nix tooling. Conceptually, one may argue that such changes need
to be locally defined where they are refered to anyways, so the
emacs-overlay does in that sense belong in the home-config where it is
being used.